Delete jar files for current extensions from observation-manager and ddisable external load extensions by now
