+++ observation-manager/observation-manager-app/src/main/java/de/lehmannet/om/ui/extension/ExtensionLoader.java
@@ -226,7 +226,6 @@ public class ExtensionLoader {
         }
         // @formatter:on
 
-
     }
 
     private void addInternalExtension(IExtension extension) {
