+++ observation-manager/observation-manager-app/src/main/java/de/lehmannet/om/ui/extension/ExternalExtensionLoader.java
@@ -15,9 +15,12 @@ import java.net.URLClassLoader;
 import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.Enumeration;
+import java.util.Iterator;
 import java.util.List;
 import java.util.ListIterator;
+import java.util.Locale;
 import java.util.Properties;
+import java.util.StringTokenizer;
 import java.util.zip.ZipEntry;
 import java.util.zip.ZipFile;
 
@@ -25,6 +28,7 @@ import javax.xml.parsers.DocumentBuilder;
 import javax.xml.parsers.DocumentBuilderFactory;
 import javax.xml.parsers.ParserConfigurationException;
 
+import org.apache.commons.lang3.StringUtils;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.w3c.dom.Document;
@@ -38,6 +42,27 @@ import de.lehmannet.om.util.ConfigLoader;
 import de.lehmannet.om.util.SchemaLoader;
 
 public class ExternalExtensionLoader {
+
+    /**
+     * Property to define extesions dir.
+     */
+    public static final String EXTENSIONS_DIR_PROPERTY = "extensions.dir";
+
+    // ---------
+    // Constants ---------------------------------------------------------
+    // ---------
+    // Extension of config files
+    private static final String MANIFEST_FILENAME = "META-INF/SCHEMATYPE";
+    // Config file XSI relation entry: XSI_Type target ending
+    private static final String CONFIG_FILE_TARGETENTRY_TYPE_ENDING = "XSI_Relation_Type";
+    // Config file XSI relation entry: XSI_Type target ending
+    private static final String CONFIG_FILE_TARGETENTRY_CLASSNAME_ENDING = "XSI_Relation_Class";
+    // Config file XSI relation entry: XSI_Type finding ending
+    private static final String CONFIG_FILE_FINDINGENTRY_TYPE_ENDING = "XSI_Relation_Finding_Type";
+    // Config file XSI relation entry: XSI_Type finding ending
+    private static final String CONFIG_FILE_FINDINGENTRY_CLASSNAME_ENDING = "XSI_Relation_Finding_Class";
+
+    private static final Object LOCK = new Object();
 
     // File name of extension config file
     private static final String EXTENSION_FILENAME = "META-INF/OM_EXTENSION";
@@ -58,9 +83,9 @@ public class ExternalExtensionLoader {
 
     public IExtension addExtension(ZipFile extension) {
 
-        List<File> jars = this.getJarFilesInExtensionFile(extension);      
+        List<File> jars = this.getJarFilesInExtensionFile(extension);
         List<URL> urlArray = this.addJarsToClassLoader(jars);
-        
+
         if (urlArray.isEmpty()) {
             return null;
         }
@@ -88,10 +113,10 @@ public class ExternalExtensionLoader {
     private List<File> getJarFilesInExtensionFile(ZipFile extension) {
         List<File> jars = new ArrayList<>();
 
-        Enumeration<? extends ZipEntry> enumeration = extension.entries();       
+        Enumeration<? extends ZipEntry> enumeration = extension.entries();
         File lastFile = null;
         // Unpack ZIP file
-        ZipEntry ze = null;        
+        ZipEntry ze = null;
         while (enumeration.hasMoreElements()) {
             ze = enumeration.nextElement();
             lastFile = this.unpack(extension, ze);
@@ -373,4 +398,105 @@ public class ExternalExtensionLoader {
 
     }
 
+    private void readConfigForExtension() throws Exception {
+        // Get JARs under extension path
+        String extPath = System.getProperty(EXTENSIONS_DIR_PROPERTY);
+        if (!StringUtils.isBlank(extPath)) {
+            File ext = new File(extPath);
+            if (ext.exists()) {
+                File[] jars = ext.listFiles((dir, name) -> name.toLowerCase(Locale.getDefault()).endsWith(".jar"));
+                if (jars != null) {
+                    for (File jar : jars) {
+                        LOGGER.debug("Searching configured dir extensions. File {}", jar.getName());
+                        scanJarFile(jar);
+                    }
+                }
+            }
+        }
+    }
+
+    private static void loadExternalConfig() {
+        // Get JARs from classpath
+        String sep = System.getProperty("path.separator");
+        String path = System.getProperty("java.class.path");
+        StringTokenizer tokenizer = new StringTokenizer(path, sep);
+        File token = null;
+        while (tokenizer.hasMoreTokens()) {
+            token = new File(tokenizer.nextToken());
+            if ((token.isFile()) && (token.getName().endsWith(".jar"))) {
+                LOGGER.debug("Searching classpath extensions. File {}", token.getName());
+                try {
+                    scanJarFile(token);
+                } catch (Exception e) {
+                    LOGGER.error("Ignoring file {}", token.getName(), e);
+                }
+            }
+        }
+
+    }
+
+    // TODO
+    private static void scanJarFile(File jar) throws Exception {
+
+        try (ZipFile archive = new ZipFile(jar)) {
+            Enumeration<? extends ZipEntry> enu = archive.entries();
+            while (enu.hasMoreElements()) {
+                ZipEntry entry = enu.nextElement();
+                String name = entry.getName();
+                if (name.toUpperCase(Locale.getDefault()).equals(MANIFEST_FILENAME)) {
+                    LOGGER.debug("Loading extension {}:{}", jar.getName(), name);
+                    try (InputStream in = archive.getInputStream(entry)) {
+
+                        Properties prop = new Properties();
+                        prop.load(in);
+                        addConfig(prop);
+                    } catch (IOException ioe) {
+                        throw new Exception("Error while accessing entry from JAR file. ", ioe);
+                    }
+                    // we can't do anything here
+                }
+            }
+        } catch (IOException zipEx) {
+            throw new Exception("Error while accessing JAR file. ", zipEx);
+        }
+
+    }
+
+    private static void addConfig(Properties newProperties) {
+        Iterator keys = newProperties.keySet().iterator();
+        String currentKey = null;
+        String prefix = null;
+        String target_classname = null;
+        String target_type = null;
+        String finding_classname = null;
+        String finding_type = null;
+
+        while (keys.hasNext()) {
+            currentKey = (String) keys.next();
+            // Check if key ends with target TYPE ending
+            if (currentKey.endsWith(CONFIG_FILE_TARGETENTRY_TYPE_ENDING)) {
+                // Get target TYPE value
+                target_type = newProperties.getProperty(currentKey);
+
+                // Get prefix (everything that is before our target TYPE ending)
+                prefix = currentKey.substring(0, currentKey.lastIndexOf(CONFIG_FILE_TARGETENTRY_TYPE_ENDING));
+                // Use prefix and CLASS ending to get target class property
+                target_classname = newProperties.getProperty(prefix + CONFIG_FILE_TARGETENTRY_CLASSNAME_ENDING);
+                // Use prefix and CLASS ending to get finding class property
+                finding_type = newProperties.getProperty(prefix + CONFIG_FILE_FINDINGENTRY_TYPE_ENDING);
+                // Use prefix and CLASS ending to get finding class property
+                finding_classname = newProperties.getProperty(prefix + CONFIG_FILE_FINDINGENTRY_CLASSNAME_ENDING);
+
+                // Add type and classname to our list of known types
+                synchronized (LOCK) {
+                    // TODO add to config loader
+                    // targets.put(target_type, target_classname);
+                    // findings.put(finding_type, finding_classname);
+                    // target_findings.put(target_type, finding_type);
+                }
+
+            }
+        }
+
+    }
 }
