+++ observation-manager/observation-manager-distribution/src/assembly/src.xml
@@ -7,41 +7,6 @@
 
     <fileSets>
         <fileSet>
-            <directory>../skychart-extension/target/</directory>
-            <outputDirectory>extensions</outputDirectory>
-            <includes>
-                <include>*.jar</include>
-            </includes>
-        </fileSet>
-        <fileSet>
-            <directory>../deepsky-extension/target/</directory>
-            <outputDirectory>extensions</outputDirectory>
-            <includes>
-                <include>*.jar</include>
-            </includes>
-        </fileSet>
-        <fileSet>
-            <directory>../solar-system-extension/target/</directory>
-            <outputDirectory>extensions</outputDirectory>
-            <includes>
-                <include>*.jar</include>
-            </includes>
-        </fileSet>
-        <fileSet>
-            <directory>../imager-extension/target/</directory>
-            <outputDirectory>extensions</outputDirectory>
-            <includes>
-                <include>*.jar</include>
-            </includes>
-        </fileSet>
-        <fileSet>
-            <directory>../variable-stars-extension/target/</directory>
-            <outputDirectory>extensions</outputDirectory>
-            <includes>
-                <include>*.jar</include>
-            </includes>
-        </fileSet>
-        <fileSet>
             <directory>../observation-manager-app/target/</directory>
             <outputDirectory>classpath</outputDirectory>
             <includes>
